Move getLogs inside a reusable helper
